# src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: MTG Deckbuilder API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /users:
    get:
      summary: List all users
      operationId: listUsers
      tags:
        - Users
      parameters:
        - name: pagesize
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: Number of items per page
        - name: pagenumber
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: Page number to retrieve
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      summary: Get a user by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update a user
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: User deleted
  /users/{id}/decks:
    get:
      summary: Get all decks for a user
      operationId: getUserDecks
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
        - name: pagesize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Number of items per page
        - name: pagenumber
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Page number to retrieve
      responses:
        '200':
          description: List of decks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deck'
...
  /tags:
    get:
      summary: List all tags
      operationId: listTags
      tags:
        - Tags
      parameters:
        - name: pagesize
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: Number of items per page
        - name: pagenumber
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: Page number to retrieve
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a new tag
      operationId: createTag
      tags:
        - Tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /tags/{id}:
    get:
      summary: Get a tag by ID
      operationId: getTagById
      tags:
        - Tags
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
    put:
      summary: Update a tag
      operationId: updateTag
      tags:
        - Tags
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
    delete:
      summary: Delete a tag
      operationId: deleteTag
      tags:
        - Tags
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Tag deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        username: { type: string }
        registration_date: { type: string, format: date-time }
        hashed_password: { type: string }
        email: { type: string }
        country: { type: string }
...
    Deck:
      type: object
      properties:
        id:
          type: integer
        deck_name:
          type: string
        creation_date:
          type: string
          format: date-time
        last_modification:
          type: string
          format: date-time
        deck_type:
          type: string
        description:
          type: string
        is_private:
          type: boolean
        tournament:
          type: string
        share_url:
          type: string
        parent_deck_id:
          type: integer
        format:
          type: integer
        user_id:
          type: integer
        deck_contents:
          type: array
          items:
            $ref: '#/components/schemas/CardDeck'
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
...
    CardDeck:
      type: object
      properties:
        card_id:
          type: integer
        deck_id:
          type: integer
        quantity:
          type: integer
...